requirements_dev.txt
pip>=21.3
bump2version>=1.0.1
wheel>=0.37.1
watchdog~=2.1.6
flake8>=4.0.1
tox~=3.24.5
coverage~=6.2
Sphinx~=4.3.2
twine~=3.7.1

bandit>=1.7.1
black>=21.12b0
isort>=5.10.1
mypy>=0.930
mypy-extensions>=0.4.3
pytest~=6.2.5

nibabel~=3.2.1
numpy~=1.22.0
pandas~=1.3.5
pillow>=9.0.0
pydicom~=2.2.2
pymedio>=0.1.8
scikit-image~=0.19.1
scipy~=1.7.3
SimpleITK~=2.1.1
